+++ vsort.h
@@ -1,7 +1,7 @@
 /**
  * VSort: High-Performance Sorting Algorithm for Apple Silicon
  *
- * Version 0.3.3
+ * Version 0.3.4
  *
  * @author Davide Santangelo <https://github.com/davidesantangelo>
  * @license MIT
@@ -41,7 +41,7 @@ extern "C"
 #define VSORT_VERSION_MAJOR 0
 #define VSORT_VERSION_MINOR 3
 #define VSORT_VERSION_PATCH 2
-#define VSORT_VERSION_STRING "0.3.3"
+#define VSORT_VERSION_STRING "0.3.4"
 
     /**
      * HyperSort - A revolutionary sorting algorithm